修复论坛功能

+++ src/main/java/com/androidweb/repository/ArticleRepository.java
@@ -3,6 +3,6 @@ package com.androidweb.repository;
 import com.androidweb.entity.bbs.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<Article,Integer> {
+public interface ArticleRepository extends JpaRepository<Article,Long> {
 
 }
